requirements.txt
shutil
pillow
requests
os